+++ lua/ergoterm/commandline.lua
@@ -31,8 +31,8 @@ local p = {
 ---@field sticky boolean?
 ---@field cleanup_on_success boolean?
 ---@field cleanup_on_failure boolean?
----@field open_on_success boolean?
----@field open_on_failure boolean?
+---@field show_on_success boolean?
+---@field show_on_failure boolean?
 ---@field text string?
 ---@field trailing string?
 
@@ -74,8 +74,8 @@ function M.parse(args)
       sticky = true,
       cleanup_on_success = true,
       cleanup_on_failure = true,
-      open_on_success = true,
-      open_on_failure = true
+      show_on_success = true,
+      show_on_failure = true
     }
 
     for _, part in ipairs(vim.split(args, " ")) do
@@ -239,9 +239,9 @@ M._all_options = {
 
   cleanup_on_failure = M._boolean_options,
 
-  open_on_success = M._boolean_options,
+  show_on_success = M._boolean_options,
 
-  open_on_failure = M._boolean_options
+  show_on_failure = M._boolean_options
 }
 
 M._term_new_options = {
@@ -258,8 +258,8 @@ M._term_new_options = {
   sticky = M._all_options.sticky,
   cleanup_on_success = M._all_options.cleanup_on_success,
   cleanup_on_failure = M._all_options.cleanup_on_failure,
-  open_on_success = M._all_options.open_on_success,
-  open_on_failure = M._all_options.open_on_failure
+  show_on_success = M._all_options.show_on_success,
+  show_on_failure = M._all_options.show_on_failure
 }
 
 M._term_update_options = {
@@ -274,8 +274,8 @@ M._term_update_options = {
   sticky = M._all_options.sticky,
   cleanup_on_success = M._all_options.cleanup_on_success,
   cleanup_on_failure = M._all_options.cleanup_on_failure,
-  open_on_success = M._all_options.open_on_success,
-  open_on_failure = M._all_options.open_on_failure,
+  show_on_success = M._all_options.show_on_success,
+  show_on_failure = M._all_options.show_on_failure,
 }
 
 M._term_send_options = {